requirements.txt
pyaudio
vosk
numpy
paho-mqtt